add navbar and router
