+++ package.json
@@ -8,6 +8,8 @@
     "rc-dialog": "^6.5.10",
     "react": "^15.6.1",
     "react-dom": "^15.6.1",
+    "react-router": "^4.1.1",
+    "react-router-dom": "^4.1.1",
     "react-scripts": "1.0.10",
     "react-transition-group": "^1.2.0",
     "webfontloader": "^1.6.28"
