+++ src/components/App.css
@@ -1,62 +1,10 @@
-.App {
+.main-content {
     width: 800px;
     margin: 0 auto 20px auto;
-    font-family: ScalaSans, sans-serif;
-}
-
-.title {
-    margin-bottom: 30px;
-}
-
-.generator-container {
-    text-align: center;
-    padding-top: 120px;
-}
-
-.progress-text {
-    margin-top: 0;
-    text-align: center;
-}
-
-.progress-transition-leave {
-    height: 55px;
-    opacity: 1;
-}
-
-.about-container {
-    margin-top: 20px;
-}
-
-.about {
-    float: right;
-    clear: both;
-}
-
-.progress-transition-leave.progress-transition-leave-active {
-    height: 0;
-    opacity: 0;
-    transition: opacity 500ms ease-in-out 200ms, height 500ms ease-in-out 500ms;
-}
-
-@media screen and (min-width: 768px) {
-    .options-container {
-        border-left: 1px solid #ddd;
-        padding-left: 25px;
-    }
 }
 
 @media not screen and (min-width: 768px) {
-    .App {
+    .main-content {
         width: 90%;
-    }
-
-    .generator-container {
-        text-align: center;
-        padding-top: 20px;
-        padding-bottom: 40px;
-    }
-
-    .options-container {
-        border-top: 1px solid #ddd;
     }
 }
