+++ src/index.css
@@ -1,5 +1,5 @@
 body {
   margin: 0;
   padding: 0;
-  font-family: sans-serif;
+  font-family: ScalaSans, sans-serif;
 }
